// src/app/models/race.ts
export interface Race {
[x: string]: any;
horseNameCouny: any;
  raceName: any;
    id: number;
    date: string;
    course: string;
    country: string;
    time: string;
    raceId: string;
    raceNumber:String;
    horseName: string;
    jockeyName: string;
    trainerName: string;
    speed: number;
    horseBallot: string;
    finishPos: number;
    dayValue: number;
    startTime: string;

  }